Moved HideRawFilesWhenNonRawExists from ImageService to SearchService
