+++ Application/services/interfaces/ISearchService.cs
@@ -13,6 +13,13 @@ public interface ISearchService
     public List<Image> SearchImages(ImageSearchSettings imageSearchSettings);
     
     /// <summary>
+    ///     Based on the given list, filters all RAW files that ALSO have a JPG in the list
+    /// </summary>
+    /// <param name="images"></param>
+    /// <returns></returns>
+    IEnumerable<Image> HideRawFilesWhenNonRawExists(IEnumerable<Image> images);
+    
+    /// <summary>
     ///     Based on the given settings, searches all projects with those requirements
     /// </summary>
     /// <param name="projectSearchSettings">The settings for the search</param>
